rename: 변수명, 파일 구조 변경
