+++ src/main/java/com/shinhan/bullsandbears/report/Report.java
@@ -1,10 +1,14 @@
-package com.shinhan.bullsandbears.domain;
+package com.shinhan.bullsandbears.report;
 
+import com.shinhan.bullsandbears.domain.Duration;
+import com.shinhan.bullsandbears.domain.StockReportHistory;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -18,11 +22,20 @@ public class Report {
   @Column(name = "report_id")
   private Long id;
 
+  @Enumerated(EnumType.STRING)
   private Duration duration;
-  private Integer amount;
+
+  private BigDecimal amount;
   private LocalDate createdAt;
 
   @OneToMany(mappedBy = "report", cascade = CascadeType.ALL, orphanRemoval = true)
-  private List<UserReportHistory> userReportHistoryList = new ArrayList<>();
+  private List<StockReportHistory> stockReportHistoryList = new ArrayList<>();
+
+  @Builder
+  public Report (Duration duration, BigDecimal amount, LocalDate createdAt){
+    this.duration = duration;
+    this.amount = amount;
+    this.createdAt = createdAt;
+  }
 
 }
